backend/requirements.txt
asgiref==3.10.0
boto3==1.42.5
certifi==2025.10.5
dj-database-url==2.2.0
Django==5.1.4
django-cors-headers==4.5.0
django-filter==24.3
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.2.1
reportlab==4.2.5
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.7.0
